+++ app-extension/src/boot/api.js
@@ -2,11 +2,62 @@ import { camelizeKeys, decamelizeKeys } from 'humps'
 
 import asteroidConfig from 'asteroid-config'
 
-export default async ({ app }) => {
+/**
+ * @private
+ */
+function getBaseURL () {
+  // Permite sobrescrever a URL base da API via query string, útil para ambiente de preview.
+  const previewURL = new URLSearchParams(window.location.search).get('$base')
+
+  if (previewURL) {
+    try {
+      const { hostname, protocol } = new URL(previewURL)
+
+      // Somente permite URLs seguras (https) e do domínio nave.dev.br para evitar possíveis ataques de SSRF.
+      if (protocol === 'https:' && hostname.endsWith('.nave.dev.br')) return previewURL
+    } catch {}
+  }
+
+  // Caso não haja uma URL de preview válida, utiliza a URL base definida nas variáveis de ambiente ou locahost.
+  return process.env.SERVER_BASE_URL || '/'
+}
+
+export default async ({ app, router }) => {
   const api = app.config.globalProperties.$axios
 
   // Defaults
-  api.defaults.baseURL = process.env.SERVER_BASE_URL || '/'
+  api.defaults.baseURL = getBaseURL()
+
+  /**
+   * Permite que parâmetros de query string iniciados com $ sejam propagados para todas as rotas,
+   * útil para ambiente de preview.
+   */
+  router.beforeEach((to, from, next) => {
+    // Extrai os parâmetros de query string iniciados com $ da rota de origem (from).
+    const dollarParams = Object.fromEntries(Object.entries(from.query).filter(([key]) => key.startsWith('$')))
+
+    /**
+     * Valida se há algum parâmetro de query string iniciado com $ que esteja presente
+     * na rota de origem (from) mas ausente na rota de destino (to).
+     */
+    const hasMissing = Object.keys(dollarParams).some(key => !(key in to.query))
+
+    if (hasMissing) {
+      next({ ...to, query: { ...dollarParams, ...to.query } })
+    } else {
+      next()
+    }
+  })
+
+  api.interceptors.request.use(config => {
+    /**
+      * Remove os parâmetros de query string iniciados com $ antes de enviar a requisição para a API, pois eles
+      * são utilizados apenas para controlar a chave do back end por exemplo.
+      */
+    if (config.params) config.params = Object.fromEntries(Object.entries(config.params).filter(([key]) => !key.startsWith('$')))
+
+    return config
+  })
 
   api.defaults.timeout = asteroidConfig.api.serverTimeout